+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Changelog
 
-## 35.12.0
+## 35.13.0
 [Full Changelog](https://github.com/uktrade/directory-components/pull/326)
 ### Implemented enhancements
 - MVP - modify skip_ga360 to remove ga360 in-view javascript if true
+
+## 35.12.0
+[Full Changelog](https://github.com/uktrade/directory-components/pull/327)
+### Implemented enhancements
+- XOT-1290 - update cookie banner to gov.uk requirements
 
 ## 35.11.0
 [Full Changelog](https://github.com/uktrade/directory-components/pull/325)